requirements.txt
Django==4.2.7
djangorestframework==3.14.0
mongoengine==0.27.0
pymongo==4.6.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0